revise Inbox.tsx and add DetailPrimitives.tsx
