+++ frontend/src/pages/Inbox.tsx
@@ -1,6 +1,8 @@
 import { useEffect, useMemo, useState } from 'react'
 import { Search, X } from 'lucide-react'
+import InboxDetailPanel from '../components/inbox/InboxDetailPanel'
 import InboxSubmissionRow from '../components/inbox/InboxSubmissionRow'
+import { formatStatus, parseSnapshotList } from '../components/inbox/detailUtils'
 import type { OpsStatus, ReviewerSubmissionSnapshot } from '../types/dashboard'
 
 type InboxState =
@@ -346,485 +348,4 @@ function normalizeFilterText(value: unknown) {
 
 function safeText(value: unknown) {
   return typeof value === 'string' ? value : ''
-}
-
-function InboxDetailPanel({
-  submission
-}: {
-  submission: ReviewerSubmissionSnapshot | null
-}) {
-  if (submission === null) {
-    return (
-      <aside className="rounded-lg border border-dashed border-slate-300 bg-white px-5 py-8 text-sm text-slate-600">
-        Select a submission to open its review details.
-      </aside>
-    )
-  }
-
-  const displayName = submission.raw.full_name.trim() || 'Unnamed submission'
-
-  return (
-    <aside className="rounded-lg border border-slate-200 bg-white shadow-sm">
-      <div className="border-b border-slate-200 px-5 py-4">
-        <p className="text-xs font-semibold uppercase tracking-[0.08em] text-slate-500">
-          Selected Submission
-        </p>
-        <h2 className="mt-1 text-xl font-semibold tracking-normal text-slate-950">
-          {displayName}
-        </h2>
-      </div>
-
-      <dl className="grid gap-4 border-b border-slate-200 px-5 py-5 text-sm">
-        <DetailField label="Submission ID" value={submission.submission_id} />
-        <DetailField
-          label="Submitted"
-          value={formatSubmittedDate(submission.raw.created_at)}
-        />
-      </dl>
-
-      <WorkflowSection submission={submission} />
-      <RawSubmissionSection submission={submission} />
-      <ParsedOutputSection submission={submission} />
-      <ResolvedOutputSection submission={submission} />
-    </aside>
-  )
-}
-
-function WorkflowSection({
-  submission
-}: {
-  submission: ReviewerSubmissionSnapshot
-}) {
-  const ops = submission.ops
-  const errors = submission.errors
-  const hasOpsMetadata = hasSnapshotValue([
-    ops.tags,
-    ops.contact_tracking,
-    ops.updated_at,
-    ops.updated_by
-  ])
-
-  return (
-    <section className="border-b border-slate-200 bg-slate-50/60 px-5 py-5">
-      <SectionHeader
-        title="Workflow"
-        description="Reviewer-facing state and lightweight diagnostics from the snapshot."
-        status={formatStatus(ops.status)}
-        statusTone={getOpsStatusTone(ops.status)}
-      />
-
-      <div className="grid gap-4 text-sm">
-        <DetailField label="Notes Preview" value={ops.notes} multiline />
-
-        {hasOpsMetadata && (
-          <dl className="grid gap-4 border-t border-slate-200 pt-4 text-sm">
-            <DetailField label="Tags" value={ops.tags} />
-            <DetailField label="Contact Tracking" value={ops.contact_tracking} />
-            <DetailField
-              label="Workflow Updated"
-              value={formatSubmittedDate(ops.updated_at)}
-            />
-            <DetailField label="Updated By" value={ops.updated_by} />
-          </dl>
-        )}
-
-        <div className="border-t border-slate-200 pt-4">
-          <div className="mb-2 flex items-center justify-between gap-3">
-            <h4 className="text-xs font-semibold uppercase tracking-[0.08em] text-slate-500">
-              Error Visibility
-            </h4>
-            <span
-              className={[
-                'inline-flex rounded-full border px-2.5 py-1 text-xs font-semibold uppercase tracking-[0.08em]',
-                errors.has_error
-                  ? 'border-rose-200 bg-rose-50 text-rose-700'
-                  : 'border-slate-200 bg-white text-slate-600'
-              ].join(' ')}
-            >
-              {errors.has_error ? 'Error' : 'Clear'}
-            </span>
-          </div>
-
-          {errors.has_error ? (
-            <div className="rounded-md border border-rose-100 bg-white px-3 py-3">
-              <p className="break-words text-sm leading-5 text-slate-900">
-                {errors.latest_error_summary.trim() || 'Latest error summary not provided'}
-              </p>
-              <dl className="mt-3 grid gap-3 text-xs sm:grid-cols-2">
-                <DetailField label="Stage" value={errors.latest_error_stage} />
-                <DetailField label="Code" value={errors.latest_error_code} />
-              </dl>
-            </div>
-          ) : (
-            <p className="rounded-md border border-slate-200 bg-white px-3 py-2 text-sm leading-5 text-slate-600">
-              No current error signal.
-            </p>
-          )}
-        </div>
-      </div>
-    </section>
-  )
-}
-
-function RawSubmissionSection({
-  submission
-}: {
-  submission: ReviewerSubmissionSnapshot
-}) {
-  return (
-    <section className="px-5 py-5">
-      <div className="mb-4">
-        <h3 className="text-sm font-semibold uppercase tracking-[0.08em] text-libelle-indigo">
-          Raw Submission
-        </h3>
-        <p className="mt-1 text-xs leading-5 text-slate-500">
-          Source-of-truth user-entered values from submissions.
-        </p>
-      </div>
-
-      <dl className="grid gap-4 text-sm">
-        <DetailField label="Full Name" value={submission.raw.full_name} />
-        <DetailField label="Email" value={submission.raw.email} />
-        <DetailField label="Location" value={submission.raw.location_raw} />
-        <DetailField label="Timezone" value={submission.raw.timezone} />
-        <DetailField label="Skills" value={submission.raw.skills_raw} multiline />
-        <DetailField label="Interests" value={submission.raw.interests} multiline />
-        <DetailField
-          label="Experience Level"
-          value={submission.raw.experience_level}
-        />
-        <DetailField label="Availability" value={submission.raw.availability} multiline />
-        <DetailField label="Motivation" value={submission.raw.motivation} multiline />
-        <LinkDetailField label="LinkedIn" value={submission.raw.linkedin_url} />
-        <LinkDetailField label="GitHub" value={submission.raw.github_url} />
-        <DetailField label="Consent Given" value={submission.raw.consent_given} />
-        <DetailField label="Resume Filename" value={submission.raw.resume_filename} />
-        <DetailField label="Resume Status" value={submission.raw.resume_status} />
-      </dl>
-    </section>
-  )
-}
-
-function ParsedOutputSection({
-  submission
-}: {
-  submission: ReviewerSubmissionSnapshot
-}) {
-  const parsed = submission.parsed
-  const isParserComplete = parsed.parser_state === 'complete'
-  const hasParserOutput = hasSnapshotValue([
-    parsed.parser_run_id,
-    parsed.created_at,
-    parsed.parser_version,
-    parsed.parsed_skills_raw,
-    parsed.parsed_location_raw,
-    parsed.parser_confidence
-  ])
-
-  return (
-    <section className="border-t border-slate-200 px-5 py-5">
-      <SectionHeader
-        title="Raw Parsed Output"
-        description="Parser-emitted values, kept separate from user-entered and resolved data."
-        status={formatStatus(parsed.parser_state)}
-        statusTone={isParserComplete ? 'success' : 'neutral'}
-      />
-
-      {!isParserComplete && !hasParserOutput ? (
-        <StateCallout tone="neutral">
-          No parser results yet. The parser has not produced a row for this submission.
-        </StateCallout>
-      ) : (
-        <dl className="grid gap-4 text-sm">
-          <DetailField label="Parser Run ID" value={parsed.parser_run_id} />
-          <DetailField
-            label="Parser Created"
-            value={formatSubmittedDate(parsed.created_at)}
-          />
-          <DetailField label="Parser Version" value={parsed.parser_version} />
-          <DetailField
-            label="Parsed Skills Raw"
-            value={parsed.parsed_skills_raw}
-            multiline
-          />
-          <DetailField
-            label="Parsed Location Raw"
-            value={parsed.parsed_location_raw}
-            multiline
-          />
-          <DetailField label="Parser Confidence" value={parsed.parser_confidence} />
-        </dl>
-      )}
-    </section>
-  )
-}
-
-function ResolvedOutputSection({
-  submission
-}: {
-  submission: ReviewerSubmissionSnapshot
-}) {
-  const resolved = submission.resolved
-  const resolvedSkillIds = parseSnapshotList(resolved.resolved_skill_ids)
-  const unknownSkills = parseSnapshotList(resolved.unknown_skills)
-  const hasUnknowns = unknownSkills.length > 0
-  const hasResolverRun = resolved.resolver_state !== 'not_run'
-
-  return (
-    <section className="border-t border-slate-200 px-5 py-5">
-      <SectionHeader
-        title="Resolved Canonical Output"
-        description="Resolver/canonical values derived from parser output. Raw parsed values are not replaced here."
-        status={formatStatus(resolved.resolver_state)}
-        statusTone={getResolverStatusTone(resolved.resolver_state)}
-      />
-
-      {!hasResolverRun ? (
-        <StateCallout tone="neutral">
-          Resolver has not run for this parser result yet.
-        </StateCallout>
-      ) : (
-        <div className="grid gap-4">
-          {resolved.resolver_state === 'zero_matches' && (
-            <StateCallout tone="warning">
-              Resolver ran but found zero canonical skill matches.
-            </StateCallout>
-          )}
-
-          {resolved.resolver_state === 'resolved' && hasUnknowns && (
-            <StateCallout tone="warning">
-              Partial resolution: some parsed values resolved, while unknown fields remain.
-            </StateCallout>
-          )}
-
-          <dl className="grid gap-4 text-sm">
-            <DetailField label="Resolver Version" value={resolved.resolver_version} />
-            <DetailField label="Aliases Version" value={resolved.aliases_version} />
-            <ListDetailField
-              label="Resolved Skill IDs"
-              values={resolvedSkillIds}
-              emptyLabel="No resolved skill IDs"
-            />
-            <ListDetailField
-              label="Unknown Skills"
-              values={unknownSkills}
-              emptyLabel="No unknown skills"
-              tone={hasUnknowns ? 'warning' : 'neutral'}
-            />
-            <DetailField
-              label="Resolver Coverage"
-              value={resolved.resolver_coverage}
-            />
-          </dl>
-        </div>
-      )}
-    </section>
-  )
-}
-
-function SectionHeader({
-  title,
-  description,
-  status,
-  statusTone
-}: {
-  title: string
-  description: string
-  status: string
-  statusTone: 'neutral' | 'success' | 'warning'
-}) {
-  return (
-    <div className="mb-4 flex flex-col gap-3 sm:flex-row sm:items-start sm:justify-between">
-      <div>
-        <h3 className="text-sm font-semibold uppercase tracking-[0.08em] text-libelle-indigo">
-          {title}
-        </h3>
-        <p className="mt-1 text-xs leading-5 text-slate-500">{description}</p>
-      </div>
-      <span
-        className={[
-          'inline-flex w-fit rounded-full border px-3 py-1 text-xs font-semibold uppercase tracking-[0.08em]',
-          statusTone === 'success'
-            ? 'border-emerald-200 bg-emerald-50 text-emerald-700'
-            : statusTone === 'warning'
-              ? 'border-amber-200 bg-amber-50 text-amber-700'
-              : 'border-slate-200 bg-slate-50 text-slate-600'
-        ].join(' ')}
-      >
-        {status}
-      </span>
-    </div>
-  )
-}
-
-function StateCallout({
-  tone,
-  children
-}: {
-  tone: 'neutral' | 'warning'
-  children: string
-}) {
-  return (
-    <p
-      className={[
-        'rounded-md border px-3 py-2 text-sm leading-5',
-        tone === 'warning'
-          ? 'border-amber-200 bg-amber-50 text-amber-800'
-          : 'border-slate-200 bg-slate-50 text-slate-600'
-      ].join(' ')}
-    >
-      {children}
-    </p>
-  )
-}
-
-function DetailField({
-  label,
-  value,
-  multiline = false
-}: {
-  label: string
-  value: string
-  multiline?: boolean
-}) {
-  return (
-    <div>
-      <dt className="text-xs font-semibold uppercase tracking-[0.08em] text-slate-500">
-        {label}
-      </dt>
-      <dd
-        className={[
-          'mt-1 break-words text-slate-900',
-          multiline ? 'whitespace-pre-wrap' : ''
-        ].join(' ')}
-      >
-        {value.trim() || 'Not provided'}
-      </dd>
-    </div>
-  )
-}
-
-function ListDetailField({
-  label,
-  values,
-  emptyLabel,
-  tone = 'neutral'
-}: {
-  label: string
-  values: string[]
-  emptyLabel: string
-  tone?: 'neutral' | 'warning'
-}) {
-  return (
-    <div>
-      <dt className="text-xs font-semibold uppercase tracking-[0.08em] text-slate-500">
-        {label}
-      </dt>
-      <dd className="mt-2">
-        {values.length > 0 ? (
-          <ul className="flex flex-wrap gap-2">
-            {values.map((value, index) => (
-              <li
-                key={`${label}-${value}-${index}`}
-                className={[
-                  'max-w-full break-words rounded-md border px-2.5 py-1 text-sm',
-                  tone === 'warning'
-                    ? 'border-amber-200 bg-amber-50 text-amber-800'
-                    : 'border-slate-200 bg-slate-50 text-slate-800'
-                ].join(' ')}
-              >
-                {value}
-              </li>
-            ))}
-          </ul>
-        ) : (
-          <span className="text-sm text-slate-500">{emptyLabel}</span>
-        )}
-      </dd>
-    </div>
-  )
-}
-
-function LinkDetailField({ label, value }: { label: string; value: string }) {
-  const href = value.trim()
-
-  if (!href) {
-    return <DetailField label={label} value="" />
-  }
-
-  return (
-    <div>
-      <dt className="text-xs font-semibold uppercase tracking-[0.08em] text-slate-500">
-        {label}
-      </dt>
-      <dd className="mt-1 break-words">
-        <a
-          className="text-libelle-indigo underline decoration-libelle-indigo/30 underline-offset-2 hover:decoration-libelle-indigo"
-          href={href}
-          rel="noreferrer"
-          target="_blank"
-        >
-          {href}
-        </a>
-      </dd>
-    </div>
-  )
-}
-
-function formatSubmittedDate(value: string) {
-  if (!value.trim()) return 'No date'
-
-  const date = new Date(value)
-  if (Number.isNaN(date.getTime())) return value
-
-  return new Intl.DateTimeFormat(undefined, {
-    month: 'short',
-    day: 'numeric',
-    year: 'numeric'
-  }).format(date)
-}
-
-function formatStatus(status: string) {
-  return status.replace(/_/g, ' ')
-}
-
-function hasSnapshotValue(values: string[]) {
-  return values.some((value) => value.trim() !== '')
-}
-
-function getResolverStatusTone(status: string) {
-  if (status === 'resolved') return 'success'
-  if (status === 'zero_matches') return 'warning'
-  return 'neutral'
-}
-
-function getOpsStatusTone(status: string) {
-  if (status === 'contacted' || status === 'in_progress') return 'success'
-  if (status === 'paused') return 'warning'
-  return 'neutral'
-}
-
-function parseSnapshotList(value: string) {
-  const text = value.trim()
-
-  if (!text) return []
-
-  try {
-    const parsed = JSON.parse(text)
-
-    if (Array.isArray(parsed)) {
-      return parsed.map(String).map((item) => item.trim()).filter(Boolean)
-    }
-  } catch {
-    return splitSnapshotList(text)
-  }
-
-  return splitSnapshotList(text)
-}
-
-function splitSnapshotList(value: string) {
-  return value
-    .split(/[,;\n]/)
-    .map((item) => item.trim())
-    .filter(Boolean)
 }
